# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/newsFeed": {
            "get": {
                "description": "Gets a list of all feeds",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a list of all feeds",
                "parameters": [
                    {
                        "type": "string",
                        "description": "source of the feed",
                        "name": "source",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "category of the feed",
                        "name": "category",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "description": "Post a query to get a list of all feeds",
                "produces": [
                    "application/json"
                ],
                "summary": "Post a query to get a list of all feeds",
                "parameters": [
                    {
                        "description": "a map of sources to list of categories",
                        "name": "query",
                        "in": "body",
                        "schema": {
                            "$ref": "#/definitions/feed.feedQuery"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/newsFeed/categories": {
            "get": {
                "description": "Gets a list of all feed categories",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a list of all feed categories",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        },
        "/newsFeed/names": {
            "get": {
                "description": "Gets a list of all feed names",
                "produces": [
                    "application/json"
                ],
                "summary": "Gets a list of all feed names",
                "responses": {
                    "200": {
                        "description": "OK"
                    }
                }
            }
        }
    },
    "definitions": {
        "feed.feedQuery": {
            "type": "object",
            "properties": {
                "query": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "array",
                        "items": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    }
}